Move event list to its own component
