+++ src/components/events/EventList/style.module.scss
@@ -1,0 +1,39 @@
+.controls {
+  align-items: center;
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  gap: 1rem;
+
+  .search {
+    background: var(--theme-elevated-background);
+    border: 1px solid var(--theme-elevated-stroke);
+    border-radius: 2rem;
+    color: inherit;
+    font-size: 1rem;
+    line-height: 1.25rem;
+    max-width: 18rem;
+    padding: 0.5rem 1rem;
+    width: 100%;
+
+    &::placeholder {
+      color: var(--theme-text-on-background-3);
+    }
+  }
+
+  .filterOption {
+    background-color: var(--theme-elevated-background);
+    border: 1px solid var(--theme-elevated-stroke);
+    border-radius: 2rem;
+
+    > div {
+      padding: 0.5rem 1rem;
+
+      > select {
+        font-size: 1rem;
+        font-weight: normal;
+        line-height: 1.25rem;
+      }
+    }
+  }
+}
