+++ src/components/events/index.tsx
@@ -3,4 +3,5 @@ export { default as CheckInModal } from './CheckInModal';
 export { default as EventCard } from './EventCard';
 export { default as EventCarousel } from './EventCarousel';
 export { default as EventDisplay } from './EventDisplay';
+export { DEFAULT_FILTER_STATE, default as EventList } from './EventList';
 export { default as EventModal } from './EventModal';
