+++ src/components/onboarding/OnboardingScreen/style.module.scss
@@ -45,6 +45,7 @@
     display: flex;
     flex: auto;
     flex-direction: column;
+    text-align: left;
 
     .endImage {
       animation: raccoon-in 3s;
